+++ packages/gatsby-page-utils/CHANGELOG.md
@@ -3,6 +3,10 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.0.25](https://github.com/gatsbyjs/gatsby/compare/gatsby-page-utils@0.0.24...gatsby-page-utils@0.0.25) (2019-10-09)
+
+**Note:** Version bump only for package gatsby-page-utils
+
 ## [0.0.24](https://github.com/gatsbyjs/gatsby/compare/gatsby-page-utils@0.0.23...gatsby-page-utils@0.0.24) (2019-10-04)
 
 ### Bug Fixes