+++ packages/gatsby-remark-images/CHANGELOG.md
@@ -3,6 +3,10 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [3.1.26](https://github.com/gatsbyjs/gatsby/compare/gatsby-remark-images@3.1.25...gatsby-remark-images@3.1.26) (2019-10-09)
+
+**Note:** Version bump only for package gatsby-remark-images
+
 ## [3.1.25](https://github.com/gatsbyjs/gatsby/compare/gatsby-remark-images@3.1.23...gatsby-remark-images@3.1.25) (2019-09-26)
 
 **Note:** Version bump only for package gatsby-remark-images